Linked C++ to C++/CLI project and C++/CLI project to C# interface
The next step involves enabling the visibility of the C++/CLI public ref class types inside C# to allow passing of information successfully

+++ DTRQController/KalmanFilter.cpp
@@ -18,7 +18,7 @@ double KalmanFilter::Filter(double value) {
 
 	values.push_back(value);
 
-	if (values.size() > memory) {
+	if ((signed int)values.size() > memory) {
 		values.erase(values.begin());
 	}
